Moved timeline and spacevehicles data files
Timeline and unmanned space vehicles data files are now in
/data/timeline.json and /data/unmanned.json, respectively.
All images are now also hosted locally, and the Hubble diagram has been
converted to a JPG, for much faster loading.
Lastly, the Mars Odyssey caption used to say "A diagram of the Hubble
Telescope". It has now been changed to "A diagram of the Mars Odyssey".